Correção de erros

+++ src/database/migrations/20200327131615-create-recipients.js
@@ -18,7 +18,7 @@ module.exports = {
         allowNull: false,
       },
       numero: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.NUMERIC,
         allowNull: true,
       },
       complemento: {
@@ -38,7 +38,7 @@ module.exports = {
         allowNull: false,
       },
       cep: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.NUMERIC,
         allowNull: false,
       },
       created_at: {
